Restore debugging logic
